+++ plugins/unplugin/src/bundlers/rspack/create.ts
@@ -13,6 +13,7 @@ import type { PluginOptions as RawPluginOptions } from "#/@types/options";
 
 import { createLogger } from "@ammolite/integration/log";
 import { createRuntime } from "@ammolite/integration/runtime";
+import { rspackVersion } from "@rspack/core";
 import { createUnplugin } from "unplugin";
 
 import { cachePlugin } from "#/plugins/cache";
@@ -42,6 +43,10 @@ const createPlugin = (createOptions?: CreatePluginOptions): Plugin => {
         const runtime: Runtime =
             createOptions?.runtime ??
             createRuntime({
+                bundler: {
+                    name: "rspack",
+                    version: rspackVersion,
+                },
                 cwd,
                 include: options?.input?.include,
                 exclude: options?.input?.exclude,