+++ plugins/unplugin/src/bundlers/vite/create.ts
@@ -9,6 +9,7 @@ import type { PluginOptions } from "#/@types/options";
 import { createLogger } from "@ammolite/integration/log";
 import { createRuntime } from "@ammolite/integration/runtime";
 import { createUnplugin } from "unplugin";
+import { version as viteVersion } from "vite";
 
 import { cachePlugin } from "#/plugins/cache";
 import { compilePlugin } from "#/plugins/compile";
@@ -39,6 +40,10 @@ const createPlugin = (createOptions?: CreatePluginOptions): Plugin => {
         const runtime: Runtime =
             createOptions?.runtime ??
             createRuntime({
+                bundler: {
+                    name: "vite",
+                    version: viteVersion,
+                },
                 cwd,
                 include: options?.input?.include,
                 exclude: options?.input?.exclude,